Moved discovery responder into device library
