+++ ASCOM.Alpaca.Device/ASCOM.Alpaca.Device.csproj
@@ -35,6 +35,7 @@
 
   <ItemGroup>
     <ProjectReference Include="..\ASCOM.Common\ASCOM.Common.csproj" />
+    <ProjectReference Include="..\ASCOM.Tools\ASCOM.Tools.csproj" />
   </ItemGroup>
 
 </Project>
