+++ ASCOM.Alpaca.Device/Discovery/Responder.cs
